% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_stability_functions.R
\name{compute_CCM}
\alias{compute_CCM}
\title{compute_CCM}
\usage{
compute_CCM(simplex_results, lib_sizes = seq(10, 100, by = 10),
  silent = TRUE, RNGseed = 42, ..., num_cores = 1)
}
\description{
Run pairwise CCM starting from the simplex_output
}
