% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_stability_functions.R
\name{compute_dynamic_stability}
\alias{compute_dynamic_stability}
\title{compute_dynamic_stability}
\usage{
compute_dynamic_stability(block,
  results_file = here::here("output/portal_ds_results.RDS"), num_cores = 2,
  ...)
}
\arguments{
\item{block}{A data.frame containing time series for the community. Each
column is a time series of abundances.}

\item{results_file}{The location of the results to be stored on disk.}

\item{num_cores}{The number of cores to use for the parallelized run of CCM}

\item{...}{Remaining arguments are passed directly into compute_smap_coeffs}
}
\value{
none
}
\description{
Compute dynamic stability for community time series. The
  analysis has multiple components:
  (1) run simplex projection on each time series to identify the optimal
      embedding dimension
  (2) generate surrogate time series, assumming that the data have just a
      seasonal pattern
  (3) run ccm on each pairwise interaction, including the surrogate data
  (4) identify the significant interactions by comparing the CCM for the real
      time series against the calculations for the surrogate data
  (5) run S-map models for each time series, using the appropriate number of
      lags, and including the important interacting variables
  (6) extract out the s-map coefficients from the models and assemble
      matrices for the system
  (7) perform eigen-decomposition on the s-map coefficient matrices
}
