% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations-jacobian-metrics.R
\name{compute_eigen_decomp}
\alias{compute_eigen_decomp}
\title{Compute eigen-decompositions of the Jacobian matrices}
\usage{
compute_eigen_decomp(smap_matrices)
}
\arguments{
\item{smap_matrices}{A list with the Jacobian matrix (of smap-coefficients)
at each time point, resulting from \code{\link{compute_smap_matrices}}}
}
\value{
A list with two elements:
\describe{
\item{\code{values}}{a list of the eigenvalues (a vector) for each time point}
\item{\code{vectors}}{a list of the eigenvectors (a matrix, each column is an
eigenvector) for each time point}
}
}
\description{
Compute eigen-decompositions of the Jacobian matrices
}
