% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations-simplex.R
\name{compute_simplex}
\alias{compute_simplex}
\title{Run univariate analysis on each time series}
\usage{
compute_simplex(block, E_list = 1:10,
  surrogate_method = "annual_spline", num_surr = 100,
  surr_params = NULL)
}
\arguments{
\item{block}{a data.frame containing time series for the community. Each
column is a time series of abundances.}

\item{E_list}{the embedding dimension or range of embedding dimensions to
search over.}

\item{surrogate_method}{which surrogate method to use:
options are "annual_spline" or methods available in
\code{\link[rEDM:make_surrogate_data]{rEDM::make_surrogate_data()}}}

\item{num_surr}{number of surrogates to compute}

\item{surr_params}{a list of named optional arguments to be passed into the
surrogate data function}
}
\value{
A tibble with columns for the species name (taken from the original
column names), the abundance time series for each species, the output from
\code{\link[rEDM:simplex]{rEDM::simplex()}}, the best embedding dimension, as determined by
the E that minimizes MAE, and surrogate time series
}
\description{
Run simplex projection models for each time series in the
\code{block} and save the output, to determine the best embedding dimension for
each column.
}
