% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_stability_calculations.R
\name{compute_smap_matrices}
\alias{compute_smap_matrices}
\title{Generate the matrices of S-map coefficients}
\usage{
compute_smap_matrices(smap_coeffs, ccm_links)
}
\arguments{
\item{smap_coeffs}{A list of the S-map coefficients for each predictor
variable (as returned from \code{\link{compute_smap_coeffs}}}

\item{ccm_links}{A data.frame containing the significant causal links. Each
row is a causal link. The columns are:
\describe{
\item{`xmap_from`}{the column index of the predictor variable in CCM}
\item{`xmap_to`}{the column index of the predicted variable in CCM}
\item{`best_E`}{the best embedding dimension for CCM}
}}
}
\value{
A list with the matrix of smap-coefficients at each time point
}
\description{
Using the S-map coefficients, assemble the appropriate
  Jacobian matrices for each time point
}
\details{
See \code{\link{compute_smap_coeffs}} for details on the input data.
  Let the variables in the system be x^{i} with i = 1..N.
  For the S-map model predicting x^{i}_{t+1}, let the coefficient
  corresponding to variable x^{j} at lag tau be c^{tau}_{ij}.
  Then the Jacobian is the block matrix:
  J = [[ C^0 C^1 ... C^K]
       [ I   0   ... 0  ]
       [ ... ... ... 0  ]
       [ ... ... I   0  ]]
  where K is the maximum lag, and C^{tau} is the matrix formed by the values
  c^{tau}_{ij}. (Note that many of these values will be 0.)

  This function computes J at each time step.
}
