% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations-jacobian-metrics.R
\name{compute_volume_contraction}
\alias{compute_volume_contraction}
\title{Compute volume contraction of the Jacobian matrics}
\usage{
compute_volume_contraction(smap_matrices, s = NULL)
}
\arguments{
\item{smap_matrices}{A list with the Jacobian matrix (of smap-coefficients)
at each time point, resulting from \code{\link[=compute_smap_matrices]{compute_smap_matrices()}}}

\item{s}{the number of species in the system (optional parameter to restrict
the analysis just to the portions of the Jacobian that are relevant for the
forecasts)}
}
\value{
a Numeric vector of the determinants of the Jacobian
}
\description{
We define the volume contraction as the rate at which an
infinitesimal volume scales through the Jacobian.
}
\details{
See \code{\link[=compute_svd_decomp]{compute_svd_decomp()}} for details on extracting the portion of
the Jacobian used for calculating the determinant.

We do this in order to account for the low-rank of the full Jacobian, which
otherwise results in a determinant of 0.

We then compute the pseudo-determinant as |det(J^\prime %*% t(J^\prime))|.
}
