% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations-smap.R
\name{get_smap_coefficients}
\alias{get_smap_coefficients}
\title{Compute S-map coefficients for a given target variable}
\usage{
get_smap_coefficients(block, lib = c(1, NROW(block)), pred = c(1,
  NROW(block)), theta = c(seq(0, 1, by = 0.1), seq(1.5, 10, by = 0.5)))
}
\arguments{
\item{block}{the input data with time delays already generated}

\item{lib}{a 2-column matrix (or 2-element vector) where each row specifies 
the first and last *rows* of the time series to use for attractor 
reconstruction}

\item{pred}{(same format as lib), but specifying the sections of the time 
series to forecast.}

\item{theta}{the nonlinear tuning parameter (theta is only relevant if 
method == "s-map")}
}
\value{
the data.frame with the s-map coefficients
}
\description{
This function is meant to be called from within 
  \code{\link{compute_smap_coeffs}}, which also pre-generates the block so 
  that the first variable is to be predicted, and the remaining columns are 
  the causal variables and lags of the predicted variable. This function 
  searches over the values of theta for the best fit (by lowest MAE), and then 
  returns the data.frame with the s-map coefficients
}
