% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{make_surrogate_annual_spline}
\alias{make_surrogate_annual_spline}
\title{make_surrogate_annual_spline}
\usage{
make_surrogate_annual_spline(day_of_year, ts, num_surr = 100)
}
\arguments{
\item{day_of_year}{a vector of the numerical day of year, e.g. January 1 = 1,
January 2 = 2, December 31 = 365 (leap years are a bit funny, but I don't
think it should have a large effect)}

\item{ts}{the original time series}

\item{num_surr}{the number of null surrogates to generate}
}
\value{
A matrix where each column is a separate surrogate with the same
  length as `ts`.
}
\description{
Generate surrogate time series by computing a mean seasonal
  trend for each year, and shuffling the residuals. This differs from
  \code{\link[rEDM]{make_surrogate_seasonal}} in that the data are not
  sampled uniformly at the same time each year. Thus, we also ask for the
  `day_of_year` as input, in order to compute the mean seasonal trend.
}
