% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_indicator_values}
\alias{plot_indicator_values}
\alias{plot_volume_contraction,}
\alias{plot_total_variance}
\alias{plot_volume_contraction}
\title{Plot numerical metrics computed on the S-map (coefficient) matrices}
\usage{
plot_indicator_values(indicator_values, y_label = "Volume Contraction",
  line_size = 1, base_size = 16, plot_file = NULL, width = 6,
  height = NULL)

plot_volume_contraction(volume_contraction, ...)

plot_total_variance(total_variance, ...)
}
\arguments{
\item{indicator_values}{a vector of the numerical metric. It is expected
that the names are coerciable into Date format}

\item{y_label}{The label for the y-axis of the resulting plot}

\item{line_size}{the line width for the plot}

\item{base_size}{the base font size}

\item{plot_file}{the filepath to where to save the plot; if \code{NULL} (the
default), then the plot is not saved to a file}

\item{width}{width of the saved plot}

\item{height}{height of the saved plot}

\item{volume_contraction}{the volume contraction values to plot}

\item{total_variance}{the total variance values to plot}
}
\value{
A ggplot object of the plot
}
\description{
\code{\link[=plot_volume_contraction]{plot_volume_contraction()}} is a thin wrapper around
\code{\link[=plot_indicator_values]{plot_indicator_values()}}, specificying the y-axis label

\code{\link[=plot_total_variance]{plot_total_variance()}} is a thin wrapper around
\code{\link[=plot_indicator_values]{plot_indicator_values()}}, specificying the y-axis label
}
