% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sim-flour-beetle.R
\name{simulate_LPA_flour_beetles}
\alias{simulate_LPA_flour_beetles}
\title{Get a data.frame of simulated abundances (3-stage flour beetles)}
\usage{
simulate_LPA_flour_beetles(params = c(b = 10.67, u_l = 0.1955, u_a =
  0.96, c_el = 0.01647, c_ea = 0.01313, c_pa = 0.35), initial_state = c(L
  = 250, P = 5, A = 100), sample_times = seq(0, 1000), ...)
}
\arguments{
\item{params}{model parameters}

\item{initial_state}{initial conditions}

\item{sample_times}{the time values at which to make observations}

\item{...}{remaining arguments to be passed to \code{\link[deSolve]{ode}}}
}
\value{
a matrix (and `deSolve`) object with the observations (and times)
}
\description{
Simulate time series from the LPA (larvae-pupae-adults) flour 
  beetle model of Dennis et al. (2001) <\url{https://doi.org/10.1890/0012-9615(2001)071[0277:ECACDI]2.0.CO;2}>
  . The initial conditions follow the experimental setup of Dennis et al. 
  (2001) and the parameters are the maximum-likelihood estimates from Table 1
  of the same, with u_a and c_pa adjusted according to the description so as 
  to give chaotic behavior.
}
