% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{add_regime_shift_highlight}
\alias{add_regime_shift_highlight}
\title{add_regime_shift_highlight}
\usage{
add_regime_shift_highlight(my_plot, lower_date = as.Date(c("1983-11-12",
  "1990-01-06", "1998-12-22", "2009-05-23")),
  upper_date = as.Date(c("1985-03-16", "1992-04-04", "1999-11-06",
  "2011-01-05")), alpha = 0.5, fill = "grey30")
}
\arguments{
\item{my_plot}{the original ggplot object}

\item{lower_date}{a vector of the beginnings of the time spans}

\item{upper_date}{a vector of the ends of the time spans}

\item{alpha}{the transparency of the bars to add to the plot}

\item{fill}{the fill color of the bars to add to the plot}
}
\value{
A ggplot object with the bars added
}
\description{
add transparent bars to highlight specific time spans
}
