% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations-ccm.R
\name{compute_ccm}
\alias{compute_ccm}
\title{Run CCM on each pair of time series}
\usage{
compute_ccm(simplex_results, lib_sizes = seq(10, 100, by = 10),
  random_libs = TRUE, num_samples = 100, replace = TRUE,
  RNGseed = 42, silent = TRUE)
}
\arguments{
\item{simplex_results}{the output of \code{\link{compute_simplex}}}

\item{lib_sizes}{the vector of library sizes to try}

\item{random_libs}{indicates whether to use randomly sampled libs}

\item{num_samples}{is the number of random samples at each lib size (this 
parameter is ignored if random_libs is FALSE)}

\item{replace}{indicates whether to sample vectors with replacement}

\item{RNGseed}{will set a seed for the random number generator, enabling 
reproducible runs of ccm with randomly generated libraries}

\item{silent}{prevents warning messages from being printed to the R console}
}
\value{
`compute_ccm` returns a tibble with columns for the variables that we
  use in CCM, the data type (whether it's the "actual" time series or
  "surrogate"), the library size, and the results from CCM
}
\description{
`compute_ccm` runs pairwise CCM based on the simplex_output -
  using the best embedding dimension from the simplex results, and computed
  for both the real data and the surrogate data. The calculations run using
  \code{\link[furrr]{future_pmap}}. Thus, parallelization should be set by
  the user, if desired, using \code{\link[future]{plan}}, prior to running.
}
