% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_stability_calculations.R
\name{compute_eigen_decomp}
\alias{compute_eigen_decomp}
\title{Compute the eigen-decomposition of the smap matrices}
\usage{
compute_eigen_decomp(smap_matrices)
}
\arguments{
\item{smap_matrices}{A list with the matrix of smap-coefficients at each
time point \code{\link{compute_smap_matrices}}}
}
\value{
A list with two elements:
  \describe{
    \item{`values`}{a list of the eigenvalues (a vector) for each time point}
    \item{`vectors`}{a list of the eigenvectors (a matrix, each column is an
      eigenvector) for each time point}
  }
}
\description{
Compute the eigen-decomposition of the smap matrices
}
