% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{norm_rescale}
\alias{norm_rescale}
\title{linear rescaling function}
\usage{
norm_rescale(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical (default = TRUE) that affects how NAs are dealt with}
}
\value{
A rescaled vector the same length as `x`
}
\description{
rescale the input vector to mean = 0, variance = 1
}
