% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_eigenvalues}
\alias{plot_eigenvalues}
\title{plot_eigenvalues}
\usage{
plot_eigenvalues(eigenvalues, num_values = 1,
  highlight_complex = FALSE, line_size = 1, plot_file = NULL,
  width = 6, height = NULL)
}
\arguments{
\item{eigenvalues}{a list of vectors for the eigenvalues:
the number of elements in the list corresponds to the time points of the
s-map model, and each element is a vector of the eigenvalues, computed
from the matrix of the s-map coefficients at that time step}

\item{num_values}{the number of eigenvalues to plot}

\item{highlight_complex}{whether to also draw points to indicate when the
dominant eigenvalue is complex}

\item{line_size}{the line width for the plot}

\item{plot_file}{the filepath to where to save the plot; if `NULL` (the
default), then the plot is not saved to a file}

\item{width}{width of the saved plot}

\item{height}{height of the saved plot}
}
\value{
the plot object
}
\description{
Visualize the dominant eigenvalue(s) from running the S-map
  model on the community time series as part of the dynamic stability analysis
}
