% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_eigenvectors}
\alias{plot_eigenvectors}
\title{plot_eigenvectors}
\usage{
plot_eigenvectors(eigenvectors, num_values = 1, id_var = "censusdate",
  add_IPR = FALSE, palette_option = "plasma", line_size = 1,
  base_size = 16, plot_file = NULL, width = 6, height = NULL)
}
\arguments{
\item{eigenvectors}{a list of matrices for the eigenvectors:
the number of elements in the list corresponds to the time points of the
s-map model, and each element is a matrix, where the columns are the
eigenvectors, in descending order, the rows correspond to the axes of the
system}

\item{num_values}{the number of eigenvectors to plot}

\item{add_IPR}{whether to also plot the Inverse Participation Ratio, a
numerical quantity that measures how evenly the different components
contribute to the eigenvector}

\item{palette_option}{the color palette to use (see \code{\link[viridis]{viridis}}
for more info)}

\item{line_size}{the line width for the plot}

\item{base_size}{the base font size}

\item{plot_file}{the filepath to where to save the plot; if `NULL` (the
default), then the plot is not saved to a file}

\item{width}{width of the saved plot}

\item{height}{height of the saved plot}
}
\value{
A ggplot object of the eigenvector plot
}
\description{
Visualize the dominant eigenvector(s) from running the S-map
  model on the community time series as part of the dynamic stability analysis
}
